update activity card styles
