+++ src/pages/SonarFeed/ActivityRenderer.module.scss
@@ -1,0 +1,67 @@
+@import "~@santiment-network/ui/mixins.scss";
+
+.description {
+  margin-bottom: 12px;
+  display: flex;
+  color: var(--waterloo);
+
+  @include responsive('phone', 'phone-xs') {
+    flex-direction: column;
+  }
+}
+
+.link {
+  color: var(--rhino) !important;
+
+  @include responsive('phone', 'phone-xs') {
+    margin-left: 0;
+  }
+}
+
+.activityItem {
+  color: var(--rhino);
+  padding: 18px 24px;
+  background: var(--white);
+  border: 1px solid var(--porcelain);
+  box-sizing: border-box;
+  box-shadow: 0 2px 8px rgba(59, 84, 106, 0.03), 0 1px 2px rgba(59, 84, 106, 0.03);
+  border-radius: $border-radius;
+  margin-bottom: 16px;
+
+  a {
+    color: var(--jungle-green);
+
+    &:hover {
+      color: var(--jungle-green-hover);
+    }
+  }
+
+  img {
+    margin: 14px 0 0;
+    border-radius: $border-radius;
+    max-width: 100%;
+    padding: 8px;
+    border: 1px solid var(--mystic);
+  }
+
+  p {
+    @include text('body-3');
+
+    margin-bottom: 0 !important;
+    color: var(--waterloo);
+  }
+}
+
+.title {
+  @include text('body-3');
+
+  display: flex;
+  width: 100%;
+  margin: 0;
+  color: var(--waterloo);
+}
+
+.date {
+  margin-left: auto;
+  color: var(--casper);
+}
