+++ src/pages/SonarFeed/SonarFeedActivityPage.js
@@ -1,16 +1,13 @@
 import React, { useEffect } from 'react'
-import { Link } from 'react-router-dom'
 import cx from 'classnames'
 import { graphql } from 'react-apollo'
-import Markdown from 'react-markdown'
 import gql from 'graphql-tag'
 import PageLoader from '../../components/Loader/PageLoader'
 import SonarFeedRecommendations from './SonarFeedRecommendations'
-import { dateDifferenceInWords } from '../../utils/dates'
 import { getSanSonarSW } from '../Account/SettingsSonarWebPushNotifications'
-import { SIGNAL_ANCHORS } from '../../ducks/Signals/common/constants'
-import styles from './SonarFeedActivityPage.module.scss'
 import { makeVariables } from '../feed/GeneralFeed/utils'
+import ActivityRenderer from './ActivityRenderer'
+import styles from './SonarFeedActivityPage.module.scss'
 
 export const TRIGGER_ACTIVITIES_QUERY = gql`
   query signalsHistoricalActivity($limit: Int, $cursor: CursorInput) {
@@ -31,43 +28,6 @@ export const TRIGGER_ACTIVITIES_QUERY = gql`
     }
   }
 `
-
-const formatDate = dateString => {
-  return dateDifferenceInWords({
-    from: new Date(dateString)
-  })
-}
-
-export const ActivityRenderer = ({
-  activity: { triggeredAt, payload, trigger: { id: signalId, title } = {} },
-  index,
-  classes = {}
-}) => (
-  <div
-    key={triggeredAt + '_' + signalId}
-    className={cx(
-      styles.activityItem,
-      classes.activityItem,
-      index === 0 && classes.firstActivity
-    )}
-  >
-    <div className={cx(styles.description, classes.activityCustom)}>
-      <h4 className={styles.date}>
-        {formatDate(triggeredAt)} by{' '}
-        <Link
-          to={`/sonar/signal/${signalId}${SIGNAL_ANCHORS.ACTIVITIES}`}
-          className={styles.link}
-        >
-          {title}
-        </Link>
-      </h4>
-    </div>
-    <Markdown
-      source={Object.values(payload)[0]}
-      className={classes.activityMarkdown}
-    />
-  </div>
-)
 
 const SonarFeedActivityPage = ({ activities, loading, classes = {} }) => {
   if (loading) {
