+++ src/pages/feed/GeneralFeed/FeedItemRenderer/FeedItemRenderer.module.scss
@@ -15,12 +15,6 @@
 }
 
 .activityItem {
-  padding: 16px;
-  background: var(--white);
-  border: 1px solid var(--porcelain);
-  box-sizing: border-box;
-  box-shadow: 0 2px 8px rgba(59, 84, 106, 0.03), 0 1px 2px rgba(59, 84, 106, 0.03);
-  border-radius: $border-radius;
   margin-bottom: 16px;
 }
